requirements.txt
streamlit>=1.35.0
langchain>=0.2.10
langchain-community>=0.2.10
langchain-huggingface>=0.0.3
langchain-ollama>=0.1.0
langchain-openai>=0.1.20
faiss-cpu>=1.8.0
pypdf>=4.3.1
python-dotenv>=1.0.1
sentence-transformers>=3.0.1
torch>=2.2.0
